+++ test/loc.js
@@ -25,12 +25,12 @@ const validRecords = [
     testT  : ':cambridge-net.kei.com:29:\\000\\063\\000\\000\\211\\027\\055\\320\\160\\276\\025\\360\\000\\230\\215\\040:3600::\n',
   },
   {
-    name   : 'rwy04L.logan-airport.boston.',
+    name   : 'rwy04l.logan-airport.boston.',
     type   : 'LOC',
     address: '42 21 28.764 N 71 0 51.617 W -44m 2000m',
     ttl    : 3600,
-    testB  : 'rwy04L.logan-airport.boston.\t3600\tIN\tLOC\t42 21 28.764 N 71 0 51.617 W -44m 2000m\n',
-    testT  : ':rwy04L.logan-airport.boston:29:\\000\\045\\000\\000\\211\\026\\313\\074\\160\\303\\020\\337\\000\\230\\205\\120:3600::\n',
+    testB  : 'rwy04l.logan-airport.boston.\t3600\tIN\tLOC\t42 21 28.764 N 71 0 51.617 W -44m 2000m\n',
+    testT  : ':rwy04l.logan-airport.boston:29:\\000\\045\\000\\000\\211\\026\\313\\074\\160\\303\\020\\337\\000\\230\\205\\120:3600::\n',
   },
 ]